+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  - Issue #144 - AppSourceCop.json Visual Editor: "unknown" property (for "publisher" property)
  - Issue #148 - Add dot at the end of tooltip
  - Issue #149 - Missing second quote in <returns name="ReturnVariableName"> in generated documentation comment
+ - Issue #142 - Ruleset editor - add rules lookup
+ - Autocomplete drop down can display value descriptions
  - Table extension wizard added by rvanbekkum
 
  Thank you rvanbekkum for reporting issues #143, #144 and #149 and your pull request with issue #149 fixes and table extension wizard